update style on guide
