+++ wondrous-app/components/Guide/index.tsx
@@ -4,7 +4,10 @@ import { SET_USER_COMPLETED_GUIDE } from 'graphql/mutations/user';
 import { useRouter } from 'next/router';
 import { toggleHtmlOverflow } from 'utils/helpers';
 import { GET_LOGGED_IN_USER } from 'graphql/queries';
-import { NextButton, PrevButton, NavigationWrapper, LaunchButton, LaunchButtonText } from './styles';
+import GuideNextButton from 'components/Guide/GuideNextButton';
+import GuidePrevButton from 'components/Guide/GuidePrevButton';
+import GuideNavigationWrapper from 'components/Guide/GuideNavigationWrapper';
+import palette from 'theme/palette';
 import { guideConfig } from './guide';
 
 export default function OnboardingGuide({ children }) {
@@ -31,6 +34,7 @@ export default function OnboardingGuide({ children }) {
   };
 
   const styles = {
+    close: (base) => ({ ...base, color: palette.grey10 }),
     popover: (base) => ({
       ...base,
       borderRadius: '6px',
@@ -57,108 +61,34 @@ export default function OnboardingGuide({ children }) {
       styles={styles}
       components={{
         Badge: () => null,
-        Navigation: ({ nextButton, prevButton, currentStep, setIsOpen, setCurrentStep, ...props }) => {
-          const NextBtn: any = nextButton;
-          const PrevBtn: any = prevButton;
-          const currentStepConfig = steps[currentStep];
-          const hideButtons = currentStepConfig?.hideButtons;
-          const EndButton = currentStepConfig?.endButton;
-          return (
-            <NavigationWrapper>
-              {(!hideButtons || !EndButton) && (
-                <>
-                  <PrevBtn currentStep={currentStep} setIsOpen={setIsOpen} setCurrentStep={setCurrentStep} />
-                  <NextBtn currentStep={currentStep} setIsOpen={setIsOpen} setCurrentStep={setCurrentStep} />
-                </>
-              )}
-              {EndButton ? <EndButton onClick={() => setIsOpen(false)} /> : null}
-            </NavigationWrapper>
-          );
-        },
+        Navigation: ({ nextButton, prevButton, currentStep, setIsOpen, setCurrentStep }) => (
+          <GuideNavigationWrapper
+            nextButton={nextButton}
+            prevButton={prevButton}
+            currentStep={currentStep}
+            setIsOpen={setIsOpen}
+            setCurrentStep={setCurrentStep}
+          />
+        ),
       }}
-      nextButton={({ currentStep, stepsLength, setIsOpen, setCurrentStep }) => {
-        const stepsData = steps[currentStep];
-        const buttonTitle = stepsData?.nextButtonTitle || 'Next';
-        const action = () => {
-          if (currentStep === steps.length - 1) {
-            return setIsOpen(false);
-          }
-          return setCurrentStep((step) => step + 1);
-        };
-        if (stepsData?.nextAction === 'skip') {
-          return null;
-        }
-        if (stepsData.nextAction === 'finish') {
-          return (
-            <LaunchButton
-              onClick={() => {
-                if (guide?.id) {
-                  setUserCompletedGuide({
-                    variables: {
-                      guideId: guide?.id,
-                    },
-                  }).then(() => {
-                    router.push('/onboarding-dao', undefined, {
-                      shallow: true,
-                    });
-                  });
-                } else {
-                  router.push('/onboarding-dao', undefined, {
-                    shallow: true,
-                  });
-                }
-              }}
-            >
-              <LaunchButtonText>Launch a Project</LaunchButtonText>
-            </LaunchButton>
-          );
-        }
-        return (
-          <NextButton type="button" onClick={action}>
-            {buttonTitle}
-          </NextButton>
-        );
-      }}
-      prevButton={({ currentStep, stepsLength, setIsOpen, setCurrentStep }) => {
-        const stepsData = steps[currentStep];
-        const buttonTitle = stepsData?.prevButtonTitle || 'Next';
-        const action = () => {
-          if (currentStep === 0 || stepsData.prevAction === 'skip') {
-            return setIsOpen(false);
-          }
-          return setCurrentStep((step) => step - 1);
-        };
-        if (stepsData.nextAction === 'finish') {
-          return (
-            <LaunchButton
-              onClick={() => {
-                if (guide?.id) {
-                  setUserCompletedGuide({
-                    variables: {
-                      guideId: guide?.id,
-                    },
-                  }).then(() => {
-                    router.push('/explore', undefined, {
-                      shallow: true,
-                    });
-                  });
-                } else {
-                  router.push('/explore', undefined, {
-                    shallow: true,
-                  });
-                }
-              }}
-            >
-              <LaunchButtonText>Find a project</LaunchButtonText>
-            </LaunchButton>
-          );
-        }
-        return (
-          <PrevButton type="button" onClick={action}>
-            {buttonTitle}
-          </PrevButton>
-        );
-      }}
+      nextButton={({ currentStep, stepsLength, setIsOpen, setCurrentStep }) => (
+        <GuideNextButton
+          currentStep={currentStep}
+          stepsLength={stepsLength}
+          setIsOpen={setIsOpen}
+          setCurrentStep={setCurrentStep}
+          setUserCompletedGuide={setUserCompletedGuide}
+        />
+      )}
+      prevButton={({ currentStep, stepsLength, setIsOpen, setCurrentStep }) => (
+        <GuidePrevButton
+          currentStep={currentStep}
+          stepsLength={stepsLength}
+          setIsOpen={setIsOpen}
+          setCurrentStep={setCurrentStep}
+          setUserCompletedGuide={setUserCompletedGuide}
+        />
+      )}
     >
       {children}
     </TourProvider>
